gulp

+++ app/Http/Controllers/ProjectController.php
@@ -3,10 +3,10 @@
 namespace AcessoProject\Http\Controllers;
 
 
-
 use AcessoProject\Repositories\ProjectRepository;
 use AcessoProject\Services\ProjectService;
 use Illuminate\Http\Request;
+use LucaDegasperi\OAuth2Server\Facades\Authorizer;
 
 class ProjectController extends Controller
 {
@@ -22,16 +22,17 @@ class ProjectController extends Controller
     private $projectService;
 
     public function __construct(
-                                ProjectRepository $projectRepository,
-                                ProjectService $projectService)
+        ProjectRepository $projectRepository,
+        ProjectService $projectService)
     {
- ;
+
         $this->projectRepository = $projectRepository;
         $this->projectService = $projectService;
     }
+
     public function index()
     {
-        return $this->projectRepository->all();
+        return $this->projectRepository->findWhere(['owner_id' => \Authorizer::getResourceOwnerId()]);
     }
 
     public function store(Request $request)
@@ -44,22 +45,52 @@ class ProjectController extends Controller
 
     public function show($id)
     {
-
+        if ($this->checkProjectPermissions($id)  == false) {
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->projectRepository->find($id);
 
     }
 
-    public function update(Request $request,$id)
+    public function update(Request $request, $id)
     {
-
-        return $this->projectService->update($request->all(),$id);
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
+        return $this->projectService->update($request->all(), $id);
 
     }
 
     public function destroy($id)
     {
-
+        if ($this->checkProjectOwner($id) == false) {
+            return ['error' => 'Access Forbidden'];
+        }
         return $this->projectRepository->delete($id);
 
+    }
+
+    private function checkProjectOwner($projectId)
+    {
+
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->projectRepository->isOwner($projectId, $userId);
+    }
+
+    private function checkProjectMember($projectId)
+    {
+
+        $userId = \Authorizer::getResourceOwnerId();
+
+        return $this->projectRepository->hasMember($projectId, $userId);
+    }
+
+    private function checkProjectPermissions($projectId)
+    {
+        if ($this->checkProjectOwner($projectId) or $this->checkProjectMember($projectId)) {
+            return true;
+        }
+            return false;
     }
 }
